backend/requirements.txt
fastapi==0.135.3
uvicorn==0.41.0
pymongo==4.16.0
pdfplumber==0.11.9
google-genai==1.72.0
python-dotenv==1.2.2
passlib==1.7.4
bcrypt==5.0.0
pydantic==2.12.5
python-multipart==0.0.24
